Move screen analysis code to OCRHelper
